+++ planet-examples/src/main/java/org/saar/example/normalmapping/NormalMappingExample.java
@@ -27,8 +27,8 @@ import org.saar.core.renderer.deferred.DeferredRenderPassesPipeline;
 import org.saar.core.renderer.deferred.DeferredRenderingOutput;
 import org.saar.core.renderer.deferred.DeferredRenderingPath;
 import org.saar.core.renderer.renderpass.RenderPassContext;
-import org.saar.core.renderer.renderpass.shadow.ShadowsRenderPass;
-import org.saar.core.renderer.renderpass.ssao.SsaoRenderPass;
+import org.saar.core.renderer.deferred.passes.ShadowsRenderPass;
+import org.saar.core.renderer.deferred.passes.SsaoRenderPass;
 import org.saar.core.renderer.shadow.ShadowsQuality;
 import org.saar.core.renderer.shadow.ShadowsRenderNode;
 import org.saar.core.renderer.shadow.ShadowsRenderNodeGroup;
