+++ planet-examples/src/main/java/org/saar/example/reflected/ReflectionExample.java
@@ -28,8 +28,8 @@ import org.saar.core.renderer.RenderingPath;
 import org.saar.core.renderer.deferred.*;
 import org.saar.core.renderer.reflection.Reflection;
 import org.saar.core.renderer.renderpass.RenderPassContext;
-import org.saar.core.renderer.renderpass.light.LightRenderPass;
-import org.saar.core.renderer.renderpass.shadow.ShadowsRenderPass;
+import org.saar.core.renderer.deferred.passes.LightRenderPass;
+import org.saar.core.renderer.deferred.passes.ShadowsRenderPass;
 import org.saar.core.renderer.shadow.ShadowsQuality;
 import org.saar.core.renderer.shadow.ShadowsRenderNode;
 import org.saar.core.renderer.shadow.ShadowsRenderNodeGroup;
